+++ Components/EmoteComponent.cs
@@ -7,7 +7,7 @@ namespace MagnumOpus.Components
     public struct EmoteComponent
     {
         public readonly int EntityId;
-        public readonly uint ChangedTick;
+        public readonly long ChangedTick;
         public readonly Emote Emote;
 
         public EmoteComponent(int entityId, Emote emote = Emote.Stand)
